+++ test/user_store.dart
@@ -2,7 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_reactive/flutter_reactive.dart';
 
 class UserStore extends ReactiveDependency {
-  final name = 'Alice'.rt;
+  final name = 'Alice'.rx;
 
   void updateName(String newName) {
     name.value = newName;